refactor: split transport modules and move OAuth Prisma lookups
Break GraphQL mutations and MCP tool-output schemas into domain files, fold
todo REST helpers into the route module, and keep auth route handlers free of
direct Prisma access so shared boundaries stay thin.
