+++ src/features/calendar/server/calendar-export-cache.ts
@@ -1,7 +1,7 @@
 import { enqueueUserCalendarExportRebuild } from "@/features/calendar/server/calendar-export-queue";
-import { getCloudflareCalendarExportsNamespace } from "@/lib/ports/runtime";
 import { sha256Base64Url } from "@/lib/crypto/web-crypto";
 import { writeCalendarFeedCacheAnalytics } from "@/lib/metrics/analytics-engine";
+import { getCloudflareCalendarExportsNamespace } from "@/lib/ports/runtime";
 
 const USER_CALENDAR_EXPORT_CACHE_VERSION = 1;
 // Keep feeds "fresh" longer so calendar clients that poll often do not force a
