+++ src/features/catalog/server/catalog-detail-page-server.ts
@@ -11,8 +11,8 @@ import {
 } from "@/features/catalog/lib/catalog-structured-data";
 import { getCoursePage } from "@/features/catalog/server/course-page-data";
 import { getTeacherPage } from "@/features/catalog/server/teacher-page-data";
-import { runCloudflareTraceSpan } from "@/lib/ports/runtime";
 import { getViewerContext } from "@/lib/auth/viewer-context";
+import { runCloudflareTraceSpan } from "@/lib/ports/runtime";
 import {
   buildSocialMetadata,
   formatSocialMetadataMessage,
