+++ src/features/dashboard/server/compact-overview-read-model.ts
@@ -1,12 +1,12 @@
 import { loadOverviewTodoBundle } from "@/features/todos/server/todo-service";
 import { type AppLocale, DEFAULT_LOCALE } from "@/i18n/config";
-import { runCloudflareTraceSpan } from "@/lib/ports/runtime";
 import { withUserDbContext } from "@/lib/db/prisma";
 import { elapsedMs, monotonicNowMs } from "@/lib/log/observability-clock";
 import {
   type WorkspaceOverviewStage,
   writeWorkspaceOverviewStageAnalytics,
 } from "@/lib/metrics/analytics-engine";
+import { runCloudflareTraceSpan } from "@/lib/ports/runtime";
 import { parseRequiredDateInput } from "@/lib/time/date-time-from-hhmm";
 import { shanghaiDayjs } from "@/lib/time/shanghai-dayjs";
 import { formatShanghaiDate } from "@/lib/time/shanghai-format";
