+++ src/features/publications/server/publication-object-service.ts
@@ -1,10 +1,10 @@
-import { getCloudflareR2PublicationsBucket } from "@/lib/ports/runtime";
 import type {
   PublicationObjectPlanRequest,
   PublicationObjectUploadParams,
 } from "@/lib/api/schemas/request-publication-ingestion-schemas";
 import type { PublicationIngestionServicePrincipal } from "@/lib/auth/service-principal";
 import { prisma } from "@/lib/db/prisma";
+import { getCloudflareR2PublicationsBucket } from "@/lib/ports/runtime";
 import {
   publicationObjectKey,
   publicationPrincipalKey,
