+++ src/features/young/server/young-event-image-service.ts
@@ -1,5 +1,5 @@
-import { getCloudflareR2PublicationsBucket } from "@/lib/ports/runtime";
 import { prisma } from "@/lib/db/prisma";
+import { getCloudflareR2PublicationsBucket } from "@/lib/ports/runtime";
 
 const YOUNG_EVENT_IMAGE_ORIGIN = "https://young.ustc.edu.cn/login/";
 const YOUNG_EVENT_IMAGE_KEY_PREFIX = "young-events/images/";
