+++ src/lib/graphql/AGENTS.md
@@ -27,6 +27,13 @@ Explicit public GraphQL transport and schema.
 - Personal mutations use the same `feature:write` scope and per-user rate-limit
   key as REST and MCP.
 
+## Mutations layout
+
+Mutation typedefs and resolvers live under `mutations/` (`todos`, `homeworks`,
+`subscriptions`, `comments`, `uploads`, `descriptions`, `bus`, `links`,
+`shared`, `index`). `schema.ts` still imports `graphqlMutationTypeDefs` /
+`graphqlMutationResolvers` from `./mutations`.
+
 ## Contracts
 
 - Schema changes update `docs/contracts/<module>.json`, `graphql.json`, and
