--- src/lib/graphql/mutations.ts
@@ -1,1553 +1,0 @@
-import type { BusPreferencePayload } from "@/features/bus/lib/bus-types";
-import { saveBusPreference } from "@/features/bus/server/bus-service";
-import { deleteOwnCommentsBatch } from "@/features/comments/server/comment-batch-delete";
-import {
-  createComment,
-  createCommentReaction,
-  deleteCommentReaction,
-  deleteOwnComment,
-  updateOwnComment,
-} from "@/features/comments/server/comment-mutations";
-import { setDashboardLinkPinStatesBatch } from "@/features/dashboard-links/server/dashboard-link-pin-batch";
-import {
-  MAX_PINNED_LINKS,
-  resolveDashboardLinkBySlug,
-  updateDashboardLinkPinState,
-} from "@/features/dashboard-links/server/dashboard-link-service";
-import { DESCRIPTION_CONTENT_MAX_LENGTH } from "@/features/descriptions/lib/description-limits";
-import {
-  type DescriptionTargetType,
-  resolveDescriptionTargetReference,
-} from "@/features/descriptions/server/description-targets";
-import { upsertDescriptionContent } from "@/features/descriptions/server/description-upsert";
-import {
-  homeworkDescriptionInputSchema,
-  homeworkTitleSchema,
-} from "@/features/homeworks/lib/homework-schema";
-import {
-  setHomeworkCompletion,
-  setHomeworkCompletions,
-} from "@/features/homeworks/server/homework-completion";
-import { createHomeworkForSection } from "@/features/homeworks/server/homework-create";
-import { homeworkDateError } from "@/features/homeworks/server/homework-dates";
-import {
-  deleteHomework,
-  updateHomework,
-} from "@/features/homeworks/server/homework-mutations";
-import { requireHomeworkItemById } from "@/features/homeworks/server/homework-read-model";
-import { prepareHomeworkUpdate } from "@/features/homeworks/server/prepare-homework-update";
-import {
-  batchUpdateUserSectionSubscriptions,
-  setUserSectionSubscriptionByJwId,
-} from "@/features/subscriptions/server/subscriptions";
-import type { TodoPriorityValue } from "@/features/todos/lib/todo-priority";
-import {
-  deleteTodosBatch,
-  setTodoCompletionsBatch,
-} from "@/features/todos/server/todo-batch-service";
-import {
-  createTodo,
-  deleteOwnedTodo,
-  updateOwnedTodo,
-} from "@/features/todos/server/todo-service";
-import { uploadConfig } from "@/features/uploads/lib/upload-config";
-import {
-  normalizeContentType,
-  sanitizeFilename,
-} from "@/features/uploads/lib/upload-utils";
-import { UploadError } from "@/features/uploads/server/upload-quota";
-import {
-  completeOwnedUploadSession,
-  createOwnedUploadSession,
-  deleteOwnedUpload,
-  renameOwnedUpload,
-} from "@/features/uploads/server/upload-service";
-import type {
-  CommentReactionType,
-  CommentVisibility,
-} from "@/generated/prisma/client";
-import { attributionFromGraphqlPrincipal } from "@/lib/audit/principal-attribution";
-import { getAuditRequestMetadata } from "@/lib/audit/write-audit-log";
-import { hasAsciiControlCharacters } from "@/lib/text/ascii-control-characters";
-import type { GraphqlPrincipal } from "./auth";
-import type { GraphqlContext } from "./context";
-import {
-  requireGraphqlId,
-  validateOptionalGraphqlId,
-} from "./input-boundaries";
-import {
-  badMutationInput,
-  forbiddenMutation,
-  mutationNotFound,
-  serviceUnavailableMutation,
-} from "./mutation-errors";
-import { requireGraphqlMutation } from "./mutation-guard";
-import {
-  commentReactionTypeResolver,
-  commentTargetTypeResolver,
-  commentVisibilityResolver,
-  dateTimeInput,
-  normalizeCommentBody,
-  normalizeIdList,
-  normalizeTodoContent,
-  normalizeTodoTitle,
-  rejectDuplicateMutationTargets,
-  rejectExplicitNullFields,
-  requireMutationBatchSize,
-  requireMutationId,
-} from "./mutation-input";
-
-export const graphqlMutationTypeDefs = /* GraphQL */ `
-  enum DescriptionTargetType {
-    COURSE
-    SECTION
-    TEACHER
-    HOMEWORK
-  }
-
-  enum CommentVisibility {
-    PUBLIC
-    LOGGED_IN_ONLY
-  }
-
-  enum CommentReactionType {
-    UPVOTE
-    DOWNVOTE
-    HEART
-    LAUGH
-    HOORAY
-    CONFUSED
-    ROCKET
-    EYES
-  }
-
-  enum CommentTargetType {
-    COURSE
-    SECTION
-    TEACHER
-    SECTION_TEACHER
-    HOMEWORK
-  }
-
-  enum BatchMutationErrorCode {
-    NOT_FOUND
-    FORBIDDEN
-    LOCKED
-    DELETED
-    SUSPENDED
-  }
-
-  enum SectionSubscriptionBatchAction {
-    ADD
-    REMOVE
-    SET
-  }
-
-  input CreateTodoInput {
-    title: String!
-    content: String
-    priority: TodoPriority = MEDIUM
-    dueAt: DateTime
-  }
-
-  input UpdateTodoInput {
-    title: String
-    content: String
-    priority: TodoPriority
-    dueAt: DateTime
-    completed: Boolean
-  }
-
-  input TodoCompletionBatchItemInput {
-    todoId: ID!
-    completed: Boolean!
-  }
-
-  input HomeworkCompletionBatchItemInput {
-    homeworkId: ID!
-    completed: Boolean!
-  }
-
-  input UpdateSectionSubscriptionsInput {
-    action: SectionSubscriptionBatchAction!
-    codes: [String!]!
-    semesterId: Int
-  }
-
-  input CreateHomeworkInput {
-    sectionJwId: Int!
-    title: String!
-    description: String
-    isMajor: Boolean! = false
-    requiresTeam: Boolean! = false
-    publishedAt: DateTime
-    submissionStartAt: DateTime
-    submissionDueAt: DateTime
-  }
-
-  input UpdateHomeworkInput {
-    title: String
-    description: String
-    isMajor: Boolean
-    requiresTeam: Boolean
-    publishedAt: DateTime
-    submissionStartAt: DateTime
-    submissionDueAt: DateTime
-  }
-
-  input BusPreferenceInput {
-    preferredOriginCampusId: Int
-    preferredDestinationCampusId: Int
-    showDepartedTrips: Boolean!
-  }
-
-  input WorkspaceLinkPinBatchItemInput {
-    slug: String!
-    pinned: Boolean!
-  }
-
-  input CreateUploadSessionInput {
-    filename: String!
-    contentType: String
-    size: Float!
-  }
-
-  input CompleteUploadSessionInput {
-    key: String!
-    filename: String!
-    contentType: String
-  }
-
-  input UpsertDescriptionInput {
-    targetType: DescriptionTargetType!
-    targetId: ID
-    sectionJwId: Int
-    courseJwId: Int
-    teacherId: Int
-    homeworkId: ID
-    content: String!
-  }
-
-  input CreateCommentInput {
-    targetType: CommentTargetType!
-    targetId: ID
-    sectionId: ID
-    sectionJwId: Int
-    courseJwId: Int
-    teacherId: ID
-    homeworkId: ID
-    sectionTeacherId: Int
-    body: String!
-    visibility: CommentVisibility = PUBLIC
-    isAnonymous: Boolean = false
-    parentId: ID
-    attachmentIds: [ID!]
-  }
-
-  input UpdateCommentInput {
-    body: String!
-    visibility: CommentVisibility
-    isAnonymous: Boolean
-    attachmentIds: [ID!]
-  }
-
-  type TodoMutationPayload {
-    id: ID!
-  }
-
-  type BatchMutationError {
-    code: BatchMutationErrorCode!
-    message: String!
-  }
-
-  type TodoCompletionBatchResult {
-    success: Boolean!
-    todoId: ID!
-    completed: Boolean!
-    todo: Todo
-    error: BatchMutationError
-  }
-
-  type TodoCompletionBatchPayload {
-    results: [TodoCompletionBatchResult!]!
-  }
-
-  type TodoDeleteBatchResult {
-    success: Boolean!
-    id: ID!
-    error: BatchMutationError
-  }
-
-  type TodoDeleteBatchPayload {
-    results: [TodoDeleteBatchResult!]!
-  }
-
-  type DeleteMutationPayload {
-    id: ID!
-    success: Boolean!
-  }
-
-  type HomeworkCompletionMutationPayload {
-    homeworkId: ID!
-    completed: Boolean!
-    completedAt: DateTime
-  }
-
-  type HomeworkCompletionBatchResult {
-    success: Boolean!
-    homeworkId: ID!
-    completed: Boolean!
-    completedAt: DateTime
-    error: BatchMutationError
-  }
-
-  type HomeworkCompletionBatchPayload {
-    results: [HomeworkCompletionBatchResult!]!
-  }
-
-  type HomeworkMutationPayload {
-    id: ID!
-    homework: Homework!
-  }
-
-  type HomeworkDeleteMutationPayload {
-    id: ID!
-    success: Boolean!
-    alreadyDeleted: Boolean!
-  }
-
-  type SectionSubscriptionMutationPayload {
-    sectionJwId: Int!
-    subscribed: Boolean!
-  }
-
-  type SectionSubscriptionBatchPayload {
-    action: SectionSubscriptionBatchAction!
-    semesterId: Int
-    matchedCodes: [String!]!
-    unmatchedCodes: [String!]!
-    addedCount: Int!
-    removedCount: Int!
-    unchangedCount: Int!
-    total: Int!
-  }
-
-  type WorkspaceLinkPinMutationPayload {
-    slug: String!
-    pinned: Boolean!
-    pinnedSlugs: [String!]!
-    maxPinnedLinks: Int!
-  }
-
-  type WorkspaceLinkPinBatchMutationPayload {
-    pinnedSlugs: [String!]!
-    maxPinnedLinks: Int!
-  }
-
-  type BusPreferenceMutationPayload {
-    preferredOriginCampusId: Int
-    preferredDestinationCampusId: Int
-    showDepartedTrips: Boolean!
-  }
-
-  type CommentMutationPayload {
-    id: ID!
-  }
-
-  type CommentDeleteBatchResult {
-    success: Boolean!
-    id: ID!
-    error: BatchMutationError
-  }
-
-  type CommentDeleteBatchPayload {
-    results: [CommentDeleteBatchResult!]!
-  }
-
-  type UploadMutationRecord {
-    id: ID!
-    key: String!
-    filename: String!
-    size: Float!
-    createdAt: DateTime!
-  }
-
-  type UploadSessionPayload {
-    key: String!
-    url: String!
-    maxFileSizeBytes: Float!
-    quotaBytes: Float!
-    usedBytes: Float!
-  }
-
-  type UploadCompletionPayload {
-    upload: UploadMutationRecord!
-    usedBytes: Float!
-    quotaBytes: Float!
-  }
-
-  type UploadMutationPayload {
-    upload: UploadMutationRecord!
-  }
-
-  type UploadDeleteMutationPayload {
-    id: ID!
-    success: Boolean!
-    deletedSize: Float!
-  }
-
-  type DescriptionMutationPayload {
-    id: ID!
-    updated: Boolean!
-  }
-
-  type CommentReactionMutationPayload {
-    commentId: ID!
-    type: CommentReactionType!
-    active: Boolean!
-    changed: Boolean!
-  }
-
-  type Mutation {
-    todoCreate(input: CreateTodoInput!): TodoMutationPayload!
-    todoUpdate(id: ID!, input: UpdateTodoInput!): TodoMutationPayload!
-    todoDelete(id: ID!): DeleteMutationPayload!
-    todoCompletionsSet(
-      items: [TodoCompletionBatchItemInput!]!
-    ): TodoCompletionBatchPayload!
-    todosDelete(ids: [ID!]!): TodoDeleteBatchPayload!
-    homeworkCreate(input: CreateHomeworkInput!): HomeworkMutationPayload!
-    homeworkUpdate(
-      id: ID!
-      input: UpdateHomeworkInput!
-    ): HomeworkMutationPayload!
-    homeworkDelete(id: ID!): HomeworkDeleteMutationPayload!
-    homeworkCompletionSet(
-      homeworkId: ID!
-      completed: Boolean!
-    ): HomeworkCompletionMutationPayload!
-    homeworkCompletionsSet(
-      items: [HomeworkCompletionBatchItemInput!]!
-    ): HomeworkCompletionBatchPayload!
-    subscriptionAdd(jwId: Int!): SectionSubscriptionMutationPayload!
-    subscriptionRemove(jwId: Int!): SectionSubscriptionMutationPayload!
-    subscriptionsImport(
-      input: UpdateSectionSubscriptionsInput!
-    ): SectionSubscriptionBatchPayload!
-    linkPinSet(
-      slug: String!
-      pinned: Boolean!
-    ): WorkspaceLinkPinMutationPayload!
-    linkPinsSet(
-      items: [WorkspaceLinkPinBatchItemInput!]!
-    ): WorkspaceLinkPinBatchMutationPayload!
-    busPreferencesSet(
-      input: BusPreferenceInput!
-    ): BusPreferenceMutationPayload!
-    descriptionSet(
-      input: UpsertDescriptionInput!
-    ): DescriptionMutationPayload!
-    commentCreate(input: CreateCommentInput!): CommentMutationPayload!
-    commentUpdate(
-      id: ID!
-      input: UpdateCommentInput!
-    ): CommentMutationPayload!
-    commentDelete(id: ID!): DeleteMutationPayload!
-    commentsDelete(ids: [ID!]!): CommentDeleteBatchPayload!
-    commentReactionAdd(
-      commentId: ID!
-      type: CommentReactionType!
-    ): CommentReactionMutationPayload!
-    commentReactionRemove(
-      commentId: ID!
-      type: CommentReactionType!
-    ): CommentReactionMutationPayload!
-    uploadSessionCreate(
-      input: CreateUploadSessionInput!
-    ): UploadSessionPayload!
-    uploadSessionComplete(
-      input: CompleteUploadSessionInput!
-    ): UploadCompletionPayload!
-    uploadRename(id: ID!, filename: String!): UploadMutationPayload!
-    uploadDelete(id: ID!): UploadDeleteMutationPayload!
-  }
-`;
-
-type CreateTodoInput = {
-  content?: string | null;
-  dueAt?: string | null;
-  priority?: TodoPriorityValue | null;
-  title: string;
-};
-
-type UpdateTodoInput = {
-  completed?: boolean | null;
-  content?: string | null;
-  dueAt?: string | null;
-  priority?: TodoPriorityValue | null;
-  title?: string | null;
-};
-
-type TodoCompletionBatchItemInput = {
-  completed: boolean;
-  todoId: string;
-};
-
-type HomeworkCompletionBatchItemInput = {
-  completed: boolean;
-  homeworkId: string;
-};
-
-type WorkspaceLinkPinBatchItemInput = {
-  pinned: boolean;
-  slug: string;
-};
-
-type SectionSubscriptionBatchAction = "add" | "remove" | "set";
-
-type UpdateSectionSubscriptionsInput = {
-  action: SectionSubscriptionBatchAction;
-  codes: string[];
-  semesterId?: number | null;
-};
-
-type CreateHomeworkInput = {
-  description?: string | null;
-  isMajor: boolean;
-  publishedAt?: string | null;
-  requiresTeam: boolean;
-  sectionJwId: number;
-  submissionDueAt?: string | null;
-  submissionStartAt?: string | null;
-  title: string;
-};
-
-type UpdateHomeworkInput = {
-  description?: string | null;
-  isMajor?: boolean | null;
-  publishedAt?: string | null;
-  requiresTeam?: boolean | null;
-  submissionDueAt?: string | null;
-  submissionStartAt?: string | null;
-  title?: string | null;
-};
-
-type CommentTargetTypeInput =
-  (typeof commentTargetTypeResolver)[keyof typeof commentTargetTypeResolver];
-
-type CreateCommentInput = {
-  attachmentIds?: string[] | null;
-  body: string;
-  courseJwId?: number | null;
-  homeworkId?: string | null;
-  isAnonymous?: boolean | null;
-  parentId?: string | null;
-  sectionId?: string | null;
-  sectionJwId?: number | null;
-  sectionTeacherId?: number | null;
-  targetId?: string | null;
-  targetType: CommentTargetTypeInput;
-  teacherId?: string | null;
-  visibility?: CommentVisibility | null;
-};
-
-type UpdateCommentInput = {
-  attachmentIds?: string[] | null;
-  body: string;
-  isAnonymous?: boolean | null;
-  visibility?: CommentVisibility | null;
-};
-
-type UpsertDescriptionInput = {
-  content: string;
-  courseJwId?: number | null;
-  homeworkId?: string | null;
-  sectionJwId?: number | null;
-  targetId?: string | null;
-  targetType: DescriptionTargetType;
-  teacherId?: number | null;
-};
-
-type CreateUploadSessionInput = {
-  contentType?: string | null;
-  filename: string;
-  size: number;
-};
-
-type CompleteUploadSessionInput = {
-  contentType?: string | null;
-  filename: string;
-  key: string;
-};
-
-const descriptionTargetTypeResolver = {
-  COURSE: "course",
-  SECTION: "section",
-  TEACHER: "teacher",
-  HOMEWORK: "homework",
-} as const satisfies Record<string, DescriptionTargetType>;
-
-const batchMutationErrorCodeResolver = {
-  NOT_FOUND: "not_found",
-  FORBIDDEN: "forbidden",
-  LOCKED: "locked",
-  DELETED: "deleted",
-  SUSPENDED: "suspended",
-} as const;
-
-const sectionSubscriptionBatchActionResolver = {
-  ADD: "add",
-  REMOVE: "remove",
-  SET: "set",
-} as const satisfies Record<string, SectionSubscriptionBatchAction>;
-
-function graphqlCommentAuditMetadata(
-  request: Request,
-  principal: Exclude<GraphqlPrincipal, { kind: "anonymous" }>,
-) {
-  return {
-    ...getAuditRequestMetadata(request),
-    ...attributionFromGraphqlPrincipal(principal),
-    source: "graphql",
-  };
-}
-
-function handleTodoFailure(result: {
-  error: "forbidden" | "no_changes" | "not_found";
-}): never {
-  if (result.error === "not_found") mutationNotFound("Todo not found.");
-  if (result.error === "forbidden") forbiddenMutation();
-  return badMutationInput("No todo changes were provided.");
-}
-
-function handleCommentFailure(result: { error: string }): never {
-  if (
-    result.error === "not_found" ||
-    result.error === "parent_not_found" ||
-    result.error === "target_not_found"
-  ) {
-    mutationNotFound("Comment target not found.");
-  }
-  if (result.error === "forbidden") forbiddenMutation();
-  if (result.error === "locked") forbiddenMutation("Comment is locked.");
-  if (result.error === "suspended") {
-    forbiddenMutation("Comment writes are suspended.");
-  }
-  return badMutationInput("Invalid comment mutation.");
-}
-
-function handleDescriptionFailure(result: { error: string }): never {
-  if (result.error === "not_found") {
-    mutationNotFound("Description target not found.");
-  }
-  if (result.error === "invalid_target") {
-    badMutationInput("Invalid description target.");
-  }
-  if (result.error === "suspended") {
-    forbiddenMutation("Description writes are suspended.");
-  }
-  return forbiddenMutation();
-}
-
-function handleHomeworkFailure(
-  result: { error: string },
-  missingTarget: "Homework" | "Section",
-): never {
-  if (result.error === "not_found") {
-    mutationNotFound(`${missingTarget} not found.`);
-  }
-  if (result.error === "mismatch") {
-    badMutationInput("Invalid section.");
-  }
-  if (result.error === "no_changes") {
-    badMutationInput("No homework changes were provided.");
-  }
-  if (result.error === "deleted") {
-    forbiddenMutation("Homework is deleted.");
-  }
-  if (result.error === "suspended") {
-    forbiddenMutation("Homework writes are suspended.");
-  }
-  return forbiddenMutation();
-}
-
-function handleUploadFailure(result: { error: string }): never {
-  if (result.error === "not_found") {
-    mutationNotFound("Upload not found.");
-  }
-  if (result.error === "suspended") {
-    forbiddenMutation("Upload writes are suspended.");
-  }
-  if (result.error === "storage_delete_failed") {
-    serviceUnavailableMutation(
-      "Upload storage deletion failed; metadata was preserved.",
-    );
-  }
-  return forbiddenMutation();
-}
-
-function handleUploadError(error: unknown): never {
-  if (error instanceof UploadError) {
-    badMutationInput(error.code);
-  }
-  throw error;
-}
-
-function normalizeUploadFilename(value: string, maxLength?: number) {
-  const filename = value.trim();
-  if (
-    !filename ||
-    (maxLength !== undefined && filename.length > maxLength) ||
-    hasAsciiControlCharacters(filename)
-  ) {
-    badMutationInput(
-      maxLength === undefined
-        ? "filename must be non-empty and contain no control characters."
-        : `filename must contain 1-${maxLength} characters and no control characters.`,
-    );
-  }
-  return sanitizeFilename(filename);
-}
-
-function normalizeUploadSize(value: number) {
-  if (
-    !Number.isSafeInteger(value) ||
-    value <= 0 ||
-    value > uploadConfig.maxFileSizeBytes
-  ) {
-    badMutationInput(
-      `size must be an integer between 1 and ${uploadConfig.maxFileSizeBytes}.`,
-    );
-  }
-  return value;
-}
-
-function normalizeHomeworkTitle(value: string) {
-  const result = homeworkTitleSchema.safeParse(value);
-  if (!result.success) {
-    badMutationInput("title must contain 1-200 characters.");
-  }
-  return result.data;
-}
-
-function normalizeHomeworkDescription(value: string | null | undefined) {
-  const result = homeworkDescriptionInputSchema.safeParse(value);
-  if (!result.success) {
-    badMutationInput("description must not exceed 4000 characters.");
-  }
-  return result.data;
-}
-
-function normalizeBatchIds(
-  values: readonly string[],
-  label: string,
-  max: number,
-) {
-  requireMutationBatchSize(values, label, max);
-  const ids = values.map((value) => requireMutationId(value, label));
-  rejectDuplicateMutationTargets(ids, label);
-  return ids;
-}
-
-function normalizeSubscriptionBatchCodes(values: readonly string[]) {
-  if (values.length > 500) {
-    badMutationInput("codes must contain at most 500 items.");
-  }
-  const codes = values.map((value) => {
-    const code = value.trim();
-    if (!code || code.length > 64) {
-      badMutationInput("codes must contain 1-64 character values.");
-    }
-    return code;
-  });
-  rejectDuplicateMutationTargets(
-    codes.map((code) => code.toUpperCase()),
-    "codes",
-  );
-  return codes;
-}
-
-async function setSectionSubscription(
-  context: GraphqlContext,
-  jwId: number,
-  subscribed: boolean,
-) {
-  const principal = await requireGraphqlMutation(
-    context,
-    "workspace.subscription",
-  );
-  const result = await setUserSectionSubscriptionByJwId({
-    sectionJwId: requireGraphqlId(jwId, "jwId"),
-    subscribed,
-    userId: principal.userId,
-  });
-  if (!result) mutationNotFound("Section not found.");
-  return result;
-}
-
-export const graphqlMutationResolvers = {
-  BatchMutationErrorCode: batchMutationErrorCodeResolver,
-  CommentVisibility: commentVisibilityResolver,
-  CommentReactionType: commentReactionTypeResolver,
-  CommentTargetType: commentTargetTypeResolver,
-  DescriptionTargetType: descriptionTargetTypeResolver,
-  SectionSubscriptionBatchAction: sectionSubscriptionBatchActionResolver,
-  Mutation: {
-    async todoCreate(
-      _parent: unknown,
-      args: { input: CreateTodoInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(context, "workspace.todo");
-      rejectExplicitNullFields(args.input, ["priority"]);
-      const todo = await createTodo({
-        userId: principal.userId,
-        title: normalizeTodoTitle(args.input.title),
-        content: normalizeTodoContent(args.input.content),
-        priority: args.input.priority ?? undefined,
-        dueAt: dateTimeInput(args.input.dueAt),
-      });
-      return { id: todo.id };
-    },
-    async todoUpdate(
-      _parent: unknown,
-      args: { id: string; input: UpdateTodoInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(context, "workspace.todo");
-      rejectExplicitNullFields(args.input, ["title", "priority", "completed"]);
-      const hasContent = Object.hasOwn(args.input, "content");
-      const hasDueAt = Object.hasOwn(args.input, "dueAt");
-      const result = await updateOwnedTodo({
-        id: requireMutationId(args.id, "id"),
-        userId: principal.userId,
-        data: {
-          completed: args.input.completed ?? undefined,
-          content: hasContent
-            ? normalizeTodoContent(args.input.content)
-            : undefined,
-          dueAt: hasDueAt ? dateTimeInput(args.input.dueAt) : undefined,
-          hasContent,
-          hasDueAt,
-          priority: args.input.priority ?? undefined,
-          title:
-            args.input.title == null
-              ? undefined
-              : normalizeTodoTitle(args.input.title),
-        },
-      });
-      if (!result.ok) handleTodoFailure(result);
-      return { id: result.todo.id };
-    },
-    async todoDelete(
-      _parent: unknown,
-      args: { id: string },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(context, "workspace.todo");
-      const id = requireMutationId(args.id, "id");
-      const result = await deleteOwnedTodo(id, principal.userId);
-      if (!result.ok) handleTodoFailure(result);
-      return { id, success: true };
-    },
-    async todoCompletionsSet(
-      _parent: unknown,
-      args: { items: TodoCompletionBatchItemInput[] },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.todo",
-        {
-          rateLimitTier: "batch",
-        },
-      );
-      const ids = normalizeBatchIds(
-        args.items.map((item) => item.todoId),
-        "todo IDs",
-        100,
-      );
-      const results = await setTodoCompletionsBatch(
-        principal.userId,
-        args.items.map((item, index) => ({
-          completed: item.completed,
-          todoId: ids[index] as string,
-        })),
-      );
-      return { results };
-    },
-    async todosDelete(
-      _parent: unknown,
-      args: { ids: string[] },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.todo",
-        {
-          rateLimitTier: "batch",
-        },
-      );
-      const ids = normalizeBatchIds(args.ids, "todo IDs", 100);
-      const results = await deleteTodosBatch(principal.userId, ids);
-      return { results };
-    },
-    async homeworkCreate(
-      _parent: unknown,
-      args: { input: CreateHomeworkInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.section-homework",
-      );
-      const input = args.input;
-      const publishedAt = dateTimeInput(input.publishedAt) ?? null;
-      const submissionStartAt = dateTimeInput(input.submissionStartAt) ?? null;
-      const submissionDueAt = dateTimeInput(input.submissionDueAt) ?? null;
-      const dateError = homeworkDateError({
-        publishedAt,
-        submissionDueAt,
-        submissionStartAt,
-      });
-      if (dateError) badMutationInput(dateError);
-
-      const result = await createHomeworkForSection(
-        principal.userId,
-        {
-          description: normalizeHomeworkDescription(input.description),
-          isMajor: input.isMajor,
-          publishedAt,
-          requiresTeam: input.requiresTeam,
-          sectionJwId: requireGraphqlId(input.sectionJwId, "sectionJwId"),
-          submissionDueAt,
-          submissionStartAt,
-          title: normalizeHomeworkTitle(input.title),
-        },
-        {
-          ...attributionFromGraphqlPrincipal(principal),
-          requestId: getAuditRequestMetadata(context.request).requestId,
-        },
-      );
-      if (!result.ok) handleHomeworkFailure(result, "Section");
-
-      const id = result.homework.id;
-      const homework = await requireHomeworkItemById({
-        homeworkId: id,
-        locale: context.locale,
-        userId: principal.userId,
-      });
-      return { id, homework };
-    },
-    async homeworkUpdate(
-      _parent: unknown,
-      args: { id: string; input: UpdateHomeworkInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.section-homework",
-      );
-      const id = requireMutationId(args.id, "id");
-      const input = args.input;
-      rejectExplicitNullFields(input, ["title", "isMajor", "requiresTeam"]);
-
-      const hasDescription = Object.hasOwn(input, "description");
-      const hasPublishedAt = Object.hasOwn(input, "publishedAt");
-      const hasSubmissionStartAt = Object.hasOwn(input, "submissionStartAt");
-      const hasSubmissionDueAt = Object.hasOwn(input, "submissionDueAt");
-      const publishedAt = hasPublishedAt
-        ? dateTimeInput(input.publishedAt)
-        : undefined;
-      const submissionStartAt = hasSubmissionStartAt
-        ? dateTimeInput(input.submissionStartAt)
-        : undefined;
-      const submissionDueAt = hasSubmissionDueAt
-        ? dateTimeInput(input.submissionDueAt)
-        : undefined;
-      const prepared = prepareHomeworkUpdate({
-        dates: {
-          hasPublishedAt,
-          hasSubmissionDueAt,
-          hasSubmissionStartAt,
-          publishedAt,
-          submissionDueAt,
-          submissionStartAt,
-        },
-        description: hasDescription
-          ? normalizeHomeworkDescription(input.description)
-          : undefined,
-        hasDescription,
-        isMajor: input.isMajor,
-        requiresTeam: input.requiresTeam,
-        title:
-          input.title == null ? undefined : normalizeHomeworkTitle(input.title),
-        userId: principal.userId,
-      });
-      if (!prepared.ok) {
-        badMutationInput(
-          prepared.error === "no_changes"
-            ? "No homework changes were provided."
-            : prepared.message,
-        );
-      }
-
-      const result = await updateHomework({
-        audit: {
-          ...attributionFromGraphqlPrincipal(principal),
-          requestId: getAuditRequestMetadata(context.request).requestId,
-        },
-        homeworkId: id,
-        update: prepared.update,
-        userId: principal.userId,
-      });
-      if (!result.ok) handleHomeworkFailure(result, "Homework");
-
-      const homework = await requireHomeworkItemById({
-        homeworkId: id,
-        locale: context.locale,
-        userId: principal.userId,
-      });
-      return { id, homework };
-    },
-    async homeworkDelete(
-      _parent: unknown,
-      args: { id: string },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.section-homework",
-      );
-      const id = requireMutationId(args.id, "id");
-      const result = await deleteHomework({
-        audit: {
-          ...attributionFromGraphqlPrincipal(principal),
-          requestId: getAuditRequestMetadata(context.request).requestId,
-        },
-        homeworkId: id,
-        userId: principal.userId,
-      });
-      if (!result.ok) handleHomeworkFailure(result, "Homework");
-      return { id, success: true, alreadyDeleted: result.alreadyDeleted };
-    },
-    async homeworkCompletionSet(
-      _parent: unknown,
-      args: { completed: boolean; homeworkId: string },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.homework",
-      );
-      const result = await setHomeworkCompletion({
-        completed: args.completed,
-        homeworkId: requireMutationId(args.homeworkId, "homeworkId"),
-        userId: principal.userId,
-      });
-      if (!result.success) mutationNotFound("Homework not found.");
-      return result;
-    },
-    async homeworkCompletionsSet(
-      _parent: unknown,
-      args: { items: HomeworkCompletionBatchItemInput[] },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.homework",
-        {
-          rateLimitTier: "batch",
-        },
-      );
-      const homeworkIds = normalizeBatchIds(
-        args.items.map((item) => item.homeworkId),
-        "homework IDs",
-        100,
-      );
-      return setHomeworkCompletions({
-        items: args.items.map((item, index) => ({
-          completed: item.completed,
-          homeworkId: homeworkIds[index],
-        })),
-        userId: principal.userId,
-      });
-    },
-    subscriptionAdd(
-      _parent: unknown,
-      args: { jwId: number },
-      context: GraphqlContext,
-    ) {
-      return setSectionSubscription(context, args.jwId, true);
-    },
-    subscriptionRemove(
-      _parent: unknown,
-      args: { jwId: number },
-      context: GraphqlContext,
-    ) {
-      return setSectionSubscription(context, args.jwId, false);
-    },
-    async subscriptionsImport(
-      _parent: unknown,
-      args: { input: UpdateSectionSubscriptionsInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.subscription",
-        {
-          rateLimitTier: "batch",
-        },
-      );
-      const input = args.input;
-      rejectExplicitNullFields(input, ["semesterId"]);
-      const codes = normalizeSubscriptionBatchCodes(input.codes);
-      if (input.action !== "set" && codes.length === 0) {
-        badMutationInput("codes must contain at least one item.");
-      }
-      const semesterId = validateOptionalGraphqlId(
-        input.semesterId,
-        "semesterId",
-      );
-      if (input.action === "set" && semesterId === undefined) {
-        badMutationInput("semesterId is required when action is SET.");
-      }
-
-      const result = await batchUpdateUserSectionSubscriptions({
-        action: input.action,
-        codes,
-        locale: context.locale,
-        semesterId,
-        userId: principal.userId,
-      });
-      if (!result) mutationNotFound("Semester not found.");
-      return {
-        action: result.action,
-        semesterId: result.semester?.id ?? null,
-        matchedCodes: result.matchedCodes,
-        unmatchedCodes: result.unmatchedCodes,
-        addedCount: result.addedCount,
-        removedCount: result.removedCount,
-        unchangedCount: result.unchangedCount,
-        total: result.total,
-      };
-    },
-    async linkPinSet(
-      _parent: unknown,
-      args: { pinned: boolean; slug: string },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.link-pin",
-      );
-      const link = resolveDashboardLinkBySlug(args.slug);
-      if (!link) badMutationInput("Unknown dashboard link slug.");
-
-      const pinnedSlugs = await updateDashboardLinkPinState({
-        action: args.pinned ? "pin" : "unpin",
-        slug: link.slug,
-        userId: principal.userId,
-      });
-      return {
-        slug: link.slug,
-        pinned: pinnedSlugs.includes(link.slug),
-        pinnedSlugs,
-        maxPinnedLinks: MAX_PINNED_LINKS,
-      };
-    },
-    async linkPinsSet(
-      _parent: unknown,
-      args: { items: WorkspaceLinkPinBatchItemInput[] },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.link-pin",
-        {
-          rateLimitTier: "batch",
-        },
-      );
-      requireMutationBatchSize(args.items, "items", 10);
-      const items = args.items.map((item) => ({
-        action: item.pinned ? ("pin" as const) : ("unpin" as const),
-        slug: item.slug.trim(),
-      }));
-      if (items.some((item) => !item.slug)) {
-        badMutationInput("slug must be non-empty.");
-      }
-
-      const result = await setDashboardLinkPinStatesBatch({
-        items,
-        userId: principal.userId,
-      });
-      if (!result.ok) {
-        badMutationInput(`Unknown dashboard link slug: ${result.slug}.`);
-      }
-      return {
-        pinnedSlugs: result.pinnedSlugs,
-        maxPinnedLinks: MAX_PINNED_LINKS,
-      };
-    },
-    async busPreferencesSet(
-      _parent: unknown,
-      args: { input: BusPreferencePayload },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.bus-preferences",
-      );
-      const input = args.input;
-      const result = await saveBusPreference(principal.userId, {
-        preferredOriginCampusId:
-          input.preferredOriginCampusId == null
-            ? null
-            : requireGraphqlId(
-                input.preferredOriginCampusId,
-                "preferredOriginCampusId",
-              ),
-        preferredDestinationCampusId:
-          input.preferredDestinationCampusId == null
-            ? null
-            : requireGraphqlId(
-                input.preferredDestinationCampusId,
-                "preferredDestinationCampusId",
-              ),
-        showDepartedTrips: input.showDepartedTrips,
-      });
-      if (!result.ok) badMutationInput(result.error);
-      return result.preference;
-    },
-    async descriptionSet(
-      _parent: unknown,
-      args: { input: UpsertDescriptionInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.description",
-      );
-      const input = args.input;
-      rejectExplicitNullFields(input, [
-        "targetId",
-        "sectionJwId",
-        "courseJwId",
-        "teacherId",
-        "homeworkId",
-      ]);
-      if (input.content.length > DESCRIPTION_CONTENT_MAX_LENGTH) {
-        badMutationInput(
-          `content must not exceed ${DESCRIPTION_CONTENT_MAX_LENGTH} characters.`,
-        );
-      }
-      const content = input.content.trim();
-
-      const target = await resolveDescriptionTargetReference({
-        courseJwId: validateOptionalGraphqlId(input.courseJwId, "courseJwId"),
-        homeworkId:
-          input.homeworkId == null
-            ? undefined
-            : requireMutationId(input.homeworkId, "homeworkId"),
-        rawTargetId:
-          input.targetId == null
-            ? undefined
-            : requireMutationId(input.targetId, "targetId"),
-        sectionJwId: validateOptionalGraphqlId(
-          input.sectionJwId,
-          "sectionJwId",
-        ),
-        targetType: input.targetType,
-        teacherId: validateOptionalGraphqlId(input.teacherId, "teacherId"),
-        verifyExistence: true,
-      });
-      if (!target.ok) {
-        if (target.error === "target_not_found") {
-          mutationNotFound("Description target not found.");
-        }
-        badMutationInput("Invalid description target.");
-      }
-
-      const result = await upsertDescriptionContent({
-        auditMetadata: {
-          ...getAuditRequestMetadata(context.request),
-          ...attributionFromGraphqlPrincipal(principal),
-          source: "graphql",
-        },
-        content,
-        targetId: target.targetId,
-        targetType: target.targetType,
-        userId: principal.userId,
-      });
-      if (!result.ok) handleDescriptionFailure(result);
-      return { id: result.id, updated: result.updated };
-    },
-    async commentCreate(
-      _parent: unknown,
-      args: { input: CreateCommentInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.comment",
-      );
-      const input = args.input;
-      rejectExplicitNullFields(input, [
-        "targetId",
-        "sectionId",
-        "sectionJwId",
-        "courseJwId",
-        "teacherId",
-        "homeworkId",
-        "sectionTeacherId",
-        "visibility",
-        "isAnonymous",
-        "attachmentIds",
-      ]);
-      const result = await createComment({
-        attachmentIds:
-          normalizeIdList(input.attachmentIds, "attachmentIds") ?? undefined,
-        auditMetadata: graphqlCommentAuditMetadata(context.request, principal),
-        content: normalizeCommentBody(input.body),
-        courseJwId:
-          input.courseJwId == null
-            ? undefined
-            : requireGraphqlId(input.courseJwId, "courseJwId"),
-        homeworkId:
-          input.homeworkId == null
-            ? undefined
-            : requireMutationId(input.homeworkId, "homeworkId"),
-        isAnonymous: input.isAnonymous === true,
-        parentId:
-          input.parentId == null
-            ? input.parentId
-            : requireMutationId(input.parentId, "parentId"),
-        rawTargetId: input.targetId,
-        sectionId: input.sectionId,
-        sectionJwId:
-          input.sectionJwId == null
-            ? undefined
-            : requireGraphqlId(input.sectionJwId, "sectionJwId"),
-        sectionTeacherId:
-          input.sectionTeacherId == null
-            ? undefined
-            : requireGraphqlId(input.sectionTeacherId, "sectionTeacherId"),
-        targetType: input.targetType,
-        teacherId: input.teacherId,
-        userId: principal.userId,
-        visibility: input.visibility ?? "public",
-      });
-      if (!result.ok) handleCommentFailure(result);
-      return { id: result.comment.id };
-    },
-    async commentUpdate(
-      _parent: unknown,
-      args: { id: string; input: UpdateCommentInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.comment",
-      );
-      rejectExplicitNullFields(args.input, [
-        "visibility",
-        "isAnonymous",
-        "attachmentIds",
-      ]);
-      const hasAttachmentUpdate = Object.hasOwn(args.input, "attachmentIds");
-      const attachmentIds =
-        normalizeIdList(args.input.attachmentIds, "attachmentIds") ?? [];
-      const result = await updateOwnComment({
-        attachmentIds,
-        auditMetadata: graphqlCommentAuditMetadata(context.request, principal),
-        body: normalizeCommentBody(args.input.body),
-        hasAttachmentUpdate,
-        id: requireMutationId(args.id, "id"),
-        isAnonymous: args.input.isAnonymous ?? undefined,
-        userId: principal.userId,
-        visibility: args.input.visibility ?? undefined,
-      });
-      if (!result.ok) handleCommentFailure(result);
-      return { id: result.comment.id };
-    },
-    async commentDelete(
-      _parent: unknown,
-      args: { id: string },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.comment",
-      );
-      const id = requireMutationId(args.id, "id");
-      const result = await deleteOwnComment({
-        auditMetadata: graphqlCommentAuditMetadata(context.request, principal),
-        commentId: id,
-        userId: principal.userId,
-      });
-      if (!result.ok) handleCommentFailure(result);
-      return { id, success: true };
-    },
-    async commentsDelete(
-      _parent: unknown,
-      args: { ids: string[] },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.comment",
-        {
-          rateLimitTier: "batch",
-        },
-      );
-      const ids = normalizeBatchIds(args.ids, "comment IDs", 50);
-      return deleteOwnCommentsBatch({
-        auditMetadata: graphqlCommentAuditMetadata(context.request, principal),
-        ids,
-        userId: principal.userId,
-      });
-    },
-    async commentReactionAdd(
-      _parent: unknown,
-      args: { commentId: string; type: CommentReactionType },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.comment",
-      );
-      const commentId = requireMutationId(args.commentId, "commentId");
-      const result = await createCommentReaction({
-        auditMetadata: graphqlCommentAuditMetadata(context.request, principal),
-        commentId,
-        type: args.type,
-        userId: principal.userId,
-      });
-      if (!result.ok) handleCommentFailure(result);
-      return {
-        commentId,
-        type: args.type,
-        active: true,
-        changed: result.changed,
-      };
-    },
-    async commentReactionRemove(
-      _parent: unknown,
-      args: { commentId: string; type: CommentReactionType },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "community.comment",
-      );
-      const commentId = requireMutationId(args.commentId, "commentId");
-      const result = await deleteCommentReaction({
-        auditMetadata: graphqlCommentAuditMetadata(context.request, principal),
-        commentId,
-        type: args.type,
-        userId: principal.userId,
-      });
-      if (!result.ok) handleCommentFailure(result);
-      return {
-        commentId,
-        type: args.type,
-        active: false,
-        changed: result.changed,
-      };
-    },
-    async uploadSessionCreate(
-      _parent: unknown,
-      args: { input: CreateUploadSessionInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.upload",
-      );
-      rejectExplicitNullFields(args.input, ["contentType"]);
-      try {
-        const result = await createOwnedUploadSession({
-          origin: new URL(context.request.url).origin,
-          upload: {
-            contentType: normalizeContentType(args.input.contentType),
-            filename: normalizeUploadFilename(args.input.filename),
-            size: normalizeUploadSize(args.input.size),
-          },
-          userId: principal.userId,
-        });
-        if (!result.ok) handleUploadFailure(result);
-        return result.session;
-      } catch (error) {
-        return handleUploadError(error);
-      }
-    },
-    async uploadSessionComplete(
-      _parent: unknown,
-      args: { input: CompleteUploadSessionInput },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.upload",
-      );
-      rejectExplicitNullFields(args.input, ["contentType"]);
-      try {
-        const result = await completeOwnedUploadSession(principal.userId, {
-          contentType:
-            args.input.contentType == null
-              ? undefined
-              : normalizeContentType(args.input.contentType),
-          filename: normalizeUploadFilename(args.input.filename),
-          key: requireMutationId(args.input.key, "key"),
-        });
-        if (!result.ok) handleUploadFailure(result);
-        return result.completion;
-      } catch (error) {
-        return handleUploadError(error);
-      }
-    },
-    async uploadRename(
-      _parent: unknown,
-      args: { filename: string; id: string },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.upload",
-      );
-      const result = await renameOwnedUpload({
-        filename: normalizeUploadFilename(args.filename, 255),
-        id: requireMutationId(args.id, "id"),
-        userId: principal.userId,
-      });
-      if (!result.ok) handleUploadFailure(result);
-      return { upload: result.upload };
-    },
-    async uploadDelete(
-      _parent: unknown,
-      args: { id: string },
-      context: GraphqlContext,
-    ) {
-      const principal = await requireGraphqlMutation(
-        context,
-        "workspace.upload",
-      );
-      const id = requireMutationId(args.id, "id");
-      const result = await deleteOwnedUpload({
-        audit: {
-          ...getAuditRequestMetadata(context.request),
-          ...attributionFromGraphqlPrincipal(principal),
-          source: "graphql",
-        },
-        id,
-        userId: principal.userId,
-      });
-      if (!result.ok) handleUploadFailure(result);
-      return {
-        id: result.deletedId,
-        success: true,
-        deletedSize: result.deletedSize,
-      };
-    },
-  },
-};
